refactor: code structure refactor using middleware like morgan http-error and express-rate-limiter handling error and MVC patter follow
