+++ server/middleWare/serverMiddleware.js
@@ -1,0 +1,10 @@
+const cors = require('cors');
+const morgan = require('morgan');
+const express = require('express');
+
+const middleware = [
+    morgan('dev',
+    cors(),
+    express.json())
+]
+module.exports= middleware
