+++ server/package.json
@@ -2,20 +2,27 @@
   "name": "server",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "start": "node index.js",
-    "start-dev": "nodemon index.js",
+    "start": "node server.js",
+    "start-dev": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "body-parser": "^1.20.2",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "express-rate-limit": "^6.10.0",
+    "http-errors": "^2.0.0",
+    "jsonwebtoken": "^9.0.0",
     "mongodb": "^5.7.0",
-     "jsonwebtoken": "^9.0.0"
+    "mongoose": "^7.5.0"
+  },
+  "devDependencies": {
+    "morgan": "^1.10.0"
   }
 }
