--- server/utilities/getData.js
@@ -1,6 +1,0 @@
-const getData = async (collection) => {
-  const data = await collection.find({}).toArray();
-  return data;
-};
-
-module.exports = getData;
