--- server/utilities/postData.js
@@ -1,6 +1,0 @@
-const postData = async (collection, data) => {
-  const data = await collection.insertOne(data);
-  return data;
-};
-
-module.exports = postData;
